move Message into separate file
